temp fix for NaN in rpm solution; ask andy to fix
